backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Vector Database & Embeddings
chromadb==0.5.20
sentence-transformers==3.3.1
torch==2.5.1

# Document Parsing
pymupdf==1.25.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
markdown==3.7
python-frontmatter==1.1.0
pygments==2.18.0

# API Enhancement (optional)
openai==1.57.2
anthropic==0.40.0
ollama==0.4.5

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
requests==2.32.3

# OSINT Tools
dnspython==2.7.0
beautifulsoup4==4.12.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
